Update start.sh
Home dir updated, as this is the default path with `git clone` directly from the mochimo github repo. And added `$HOME` in place of `/home/user` as it will default to the current user and should work universally for everyone. As always, test as needed and only pull if it's a valid change.